apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "benchloop-api"
version = "0.1.0"
description = "Benchloop FastAPI product core"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13,<2.0",
  "fastapi[standard]>=0.115,<1.0",
  "httpx>=0.28,<1",
  "pydantic-settings>=2.2,<3.0",
  "pyjwt[crypto]>=2.9,<3.0",
  "psycopg>=3.1,<4.0",
  "sqlalchemy>=2.0,<3.0",
]

[dependency-groups]
dev = [
  "pytest>=8,<9",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
